Added: more region-selection shapes
Apart from rectangular region-selection, the following shapes are now
also available: square, ellipse, circle, fence (point-to-point) and
lasso.
Any of these can be chosen from a dropdown box on the new Selection
toolbar.
Also part of the new toolbar is a button to make region-selection the
default selection method (i.e. no need to hold down Alt, which is then
used for regular single-click selection). The hotkey for this toggle is
Alt-S. The last button on the toolbar restricts selection to objects
that are fully enclosed within the selection shape.
Also new is the Select menu, which allows to select all, select none or
invert the current selection.