Added: selection dialog
The new selection dialog allows selecting objects by name (it is not
available when editing subobjects or UVs).
It is accessible through the new "Select by name" item in the Select
menu or by pressing the hotkey N.
It is also available in object-picking modes to pick one or more objects
for the intended purpose of the mode instead of selecting them.