Enhanced: toolbar system
Instead of rows of toolbar bundles, there are now bundles of toolbar
rows. This means that it is no longer possible to place multiple toolbar
bundles side-by-side on the same row (which could waste a lot of space
if the bundles consisted of very differently-sized toolbars), but a
bundle can now contain multiple rows of toolbars that do not leave any
space between them, leading to a much tighter layout.
Furthermore, to the right of each toolbar row exists a new type of
widget: a toolbar row handle. This handle has a grip at its left-hand
side, making it possible to drag an entire row of toolbars at once.
To create a bundle, it now suffices to drag and drop either a single
toolbar or a row of toolbars onto a row handle. This will add yet
another new kind of widget to that row handle: a bundle handle. This one
also has its own grip (at its right-hand side) allowing the entire
bundle to be moved to another location as well.
To move a toolbar (or row) out of its row (or bundle), simply drag it by
its grip to the desired location until the associated insertion marker
shows up.
Mousing over a toolbar grip now shows the name of the toolbar in a
tooltip.