Fix empty export
It is no longer possible to export an empty scene.
When nothing is selected, the user will have the opportunity to export the entire scene, but only if it contains any objects.

Also fixed a problem with dialog buttons not being disabled when an input field is active (could lead to a crash with certain dialogs).