Added: (sub)object snapping
Using the new Snap and Snap Options buttons on the Snap/Align toolbar, the  (sub)object selection can be made to snap to an increment or from one specific point (grid point, object center or pivot, or subobject center) to another, while transforming.

Additionally, the transform center, as well as the reference coordinate origin, can be precisely positioned at a specific point also.