Added: alternative interactive rotation methods
Apart from the default circular dragging of the mouse in the plane of rotation, it is now also possible to do this in the view plane, or to drag along a straight line.
These new methods are especially helpful when the view is more or less perpendicular to the plane of rotation.
They can be configured using the new Transform options dialog (Options menu -> Tools -> Transforms).