Enhanced: text to right of checkbox/radiobox integrated into widget
The text describing the functionality of a checkbox or radiobox is now an integral part of that widget instead of a separate widget. This makes these controls easier to use (bigger click area), more intuitive (color of text indicates whether or not the widget is disabled) and more efficient (less sizers are needed, which in turn leads to simpler code).