Added: UV editing of model primitives
Due to the new way model primitives are implemented since the release of version 1.0.0, it was no longer possible to edit their UVs.
This commit brings back the ability to do so, albeit in a slightly different way. Instead of having access to UV vertices, edges or polys, now the distinct parts of a primitive can be edited in UV space.